feat: TSK-010 to TSK-013, HOTFIX-tgid, Profile & BalanceManager
Backend:
- Fix telegram_id BigInt support (HOTFIX-tgid)
- Add alembic migration 003_tgid_bigint
- Update user/order models for BigInt
- Add telegram bot __main__ entry point
- Add e2e_seed.py for testing

Frontend:
- Add Profile page with user info and balance
- Add BalanceManager component for managers
- Update manager page with balance management
- Update FortuneWheel with improved UX
- Update API hooks and types

Docs:
- Update frontend-components.md with new components
- Update roles.md documentation
- Add DOCKER_ENV_GUIDE.md

Tasks completed: TSK-010, TSK-011, TSK-012, TSK-013

+++ backend/src/models/order.py
@@ -1,7 +1,7 @@
 from datetime import date, datetime
 from decimal import Decimal
 
-from sqlalchemy import JSON, Date, DateTime, ForeignKey, Integer, Numeric, String, Text, func
+from sqlalchemy import JSON, BigInteger, Date, DateTime, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .base import Base
@@ -11,7 +11,7 @@ class Order(Base):
     __tablename__ = "orders"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    user_tgid: Mapped[int] = mapped_column(Integer, ForeignKey("users.tgid"), nullable=False)
+    user_tgid: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.tgid"), nullable=False)
     cafe_id: Mapped[int] = mapped_column(Integer, ForeignKey("cafes.id"), nullable=False)
     order_date: Mapped[date] = mapped_column(Date, nullable=False)
     status: Mapped[str] = mapped_column(String(50), default="pending", nullable=False)
